Make State's terminal member more specific
Now it's an integer as opposed to a boolean, so it knows not only when
it's a terminal state but whether it's a successful termination or not. I
also changed the pos parameter of _transition_to_new_state to new_pos, so
that I don't get anything meaningful if I forget the self in self.pos.