Print performance during training less frequently
Note that the changes to the hyperparameters don't seem to have helped
things. In fact they seem to have hurt things, but that's probably just
because epsilon stays higher longer so of course it doesn't get as far.
This leads me to believe that I need to test performance using a separate
run with epsilon = 0.