Be smarter about avoiding repeated actions
Now when the training agent tries to choose a repeat action, it's forced
to select a random untested one instead. (If the training agent enters a
state in which all actions have been tried, it chooses randomly from among
all possible actions.) This ensures that the agent does not get caught in
loops.